+++ components/camel-arangodb/src/test/java/org/apache/camel/component/arangodb/integration/ArangoGraphQueriesIT.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.arangodb;
+package org.apache.camel.component.arangodb.integration;
 
 import java.util.Collection;
 
@@ -30,7 +30,7 @@ import static org.hamcrest.MatcherAssert.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertTrue;
 
-public class ArangoGraphQueriesTest extends BaseGraphTest {
+public class ArangoGraphQueriesIT extends BaseGraph {
     @Override
     protected RouteBuilder createRouteBuilder() {
         return new RouteBuilder() {
