+++ components/camel-arangodb/src/test/java/org/apache/camel/component/arangodb/integration/BaseCollection.java
@@ -14,13 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.arangodb;
+package org.apache.camel.component.arangodb.integration;
 
 import com.arangodb.ArangoCollection;
 import org.junit.jupiter.api.AfterEach;
 import org.junit.jupiter.api.BeforeEach;
 
-public class BaseCollectionTest extends BaseArangoDbTest {
+public class BaseCollection extends BaseArangoDb {
 
     protected ArangoCollection collection;
