+++ components/camel-arangodb/src/test/java/org/apache/camel/component/arangodb/integration/TestEntity.java
@@ -14,24 +14,13 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.camel.component.arangodb;
+package org.apache.camel.component.arangodb.integration;
 
-import com.arangodb.entity.DocumentEntity;
-
-public class TestDocumentEntity extends DocumentEntity {
+public class TestEntity {
     private String foo;
-    private long number;
 
-    public TestDocumentEntity() {
-    }
-
-    public TestDocumentEntity(String foo) {
+    public TestEntity(final String foo) {
         this.foo = foo;
-    }
-
-    public TestDocumentEntity(String foo, long number) {
-        this.foo = foo;
-        this.number = number;
     }
 
     public String getFoo() {
@@ -40,13 +29,5 @@ public class TestDocumentEntity extends DocumentEntity {
 
     public void setFoo(final String foo) {
         this.foo = foo;
-    }
-
-    public long getNumber() {
-        return number;
-    }
-
-    public void setNumber(long number) {
-        this.number = number;
     }
 }
